`ClassDetailScreen.kt` refactor
- Moved components into separate files
