+++ app/src/main/java/com/cornellappdev/uplift/ui/screens/ClassDetailScreen.kt
@@ -1,11 +1,12 @@
 package com.cornellappdev.uplift.ui.screens
 
-import android.content.Intent
-import android.provider.CalendarContract
-import androidx.compose.foundation.*
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.Icon
 import androidx.compose.material.Surface
 import androidx.compose.material.Text
@@ -18,19 +19,17 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.core.content.ContextCompat.startActivity
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
 import coil.compose.AsyncImage
 import com.cornellappdev.uplift.R
 import com.cornellappdev.uplift.models.UpliftClass
-import com.cornellappdev.uplift.ui.components.ClassInfoCard
+import com.cornellappdev.uplift.ui.components.classdetail.*
 import com.cornellappdev.uplift.ui.viewmodels.ClassDetailViewModel
 import com.cornellappdev.uplift.ui.viewmodels.HomeViewModel
 import com.cornellappdev.uplift.util.PRIMARY_BLACK
@@ -49,7 +48,6 @@ fun ClassDetailScreen(
     onBack: () -> Unit
 ) {
     val upliftClass by classDetailViewModel.classFlow.collectAsState()
-    val context = LocalContext.current
 
     val scrollState = rememberScrollState()
 
@@ -167,186 +165,30 @@ fun ClassDetailScreen(
         }
 
         // Date & Time Information
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(Color.White),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Spacer(Modifier.height(24.dp))
-            Text(
-                text = upliftClass?.dateAsString() ?: "",
-                fontWeight = FontWeight(300),
-                fontSize = 16.sp,
-                lineHeight = 19.5.sp,
-                textAlign = TextAlign.Center,
-                color = PRIMARY_BLACK,
-                fontFamily = montserratFamily
-            )
-            Spacer(Modifier.height(8.dp))
-            Text(
-                text = upliftClass?.time.toString(),
-                fontWeight = FontWeight(500),
-                fontSize = 16.sp,
-                lineHeight = 19.5.sp,
-                textAlign = TextAlign.Center,
-                color = PRIMARY_BLACK,
-                fontFamily = montserratFamily
-            )
-            Spacer(Modifier.height(24.dp))
-            Column(
-                modifier = Modifier
-                    .clickable {
-                        upliftClass?.let {
-                            val intent = Intent(Intent.ACTION_EDIT)
-                            intent.type = "vnd.android.cursor.item/event"
-                            intent.putExtra("beginTime", it.time.start.timeInMillis(it.date))
-                            intent.putExtra("allDay", false)
-                            intent.putExtra("endTime", it.time.end.timeInMillis(it.date))
-                            intent.putExtra(CalendarContract.Events.EVENT_LOCATION, it.location)
-                            intent.putExtra("title", it.name)
-                            startActivity(context, intent, null)
-                        }
-                    },
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Icon(
-                    painter = painterResource(id = R.drawable.ic_calendar),
-                    contentDescription = "Add to calendar",
-                    tint = PRIMARY_BLACK
-                )
-                Spacer(Modifier.height(4.dp))
-                Text(
-                    text = "ADD TO CALENDAR",
-                    fontWeight = FontWeight(700),
-                    fontSize = 12.sp,
-                    lineHeight = 14.63.sp,
-                    textAlign = TextAlign.Center,
-                    color = PRIMARY_BLACK,
-                    fontFamily = montserratFamily
-                )
-            }
-            Spacer(Modifier.height(24.dp))
-        }
+        ClassDateAndTime(upliftClass = upliftClass)
 
         LineSpacer()
 
         // Function
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(Color.White),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Spacer(Modifier.height(24.dp))
-            Text(
-                text = "FUNCTION",
-                fontWeight = FontWeight(700),
-                fontSize = 16.sp,
-                lineHeight = 19.5.sp,
-                textAlign = TextAlign.Center,
-                color = PRIMARY_BLACK,
-                fontFamily = montserratFamily
-            )
-            Spacer(Modifier.height(12.dp))
-            Text(
-                text = upliftClass?.functionsAsString() ?: "",
-                fontWeight = FontWeight(300),
-                fontSize = 14.sp,
-                lineHeight = 17.07.sp,
-                textAlign = TextAlign.Center,
-                color = PRIMARY_BLACK,
-                fontFamily = montserratFamily,
-                modifier = Modifier.padding(horizontal = 48.dp)
-            )
-            Spacer(Modifier.height(24.dp))
-        }
+        ClassFunction(upliftClass = upliftClass)
 
         LineSpacer()
 
         // Preparation
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(Color.White),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Spacer(Modifier.height(24.dp))
-            Text(
-                text = "PREPARATION",
-                fontWeight = FontWeight(700),
-                fontSize = 16.sp,
-                lineHeight = 19.5.sp,
-                textAlign = TextAlign.Center,
-                color = PRIMARY_BLACK,
-                fontFamily = montserratFamily
-            )
-            Spacer(Modifier.height(12.dp))
-            Text(
-                text = upliftClass?.preparation ?: "",
-                fontWeight = FontWeight(300),
-                fontSize = 14.sp,
-                lineHeight = 17.07.sp,
-                textAlign = TextAlign.Center,
-                color = PRIMARY_BLACK,
-                fontFamily = montserratFamily,
-                modifier = Modifier.padding(horizontal = 48.dp)
-            )
-            Spacer(Modifier.height(24.dp))
-        }
+        ClassPreparation(upliftClass = upliftClass)
 
         LineSpacer()
 
         // Description
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(Color.White),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Spacer(Modifier.height(24.dp))
-            Text(
-                text = upliftClass?.description ?: "",
-                fontWeight = FontWeight(300),
-                fontSize = 14.sp,
-                lineHeight = 17.07.sp,
-                textAlign = TextAlign.Center,
-                color = PRIMARY_BLACK,
-                fontFamily = montserratFamily,
-                modifier = Modifier.padding(horizontal = 48.dp)
-            )
-            Spacer(Modifier.height(24.dp))
-        }
+        ClassDescription(upliftClass = upliftClass)
 
         LineSpacer()
 
         // Next Sessions
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(Color.White),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Spacer(Modifier.height(24.dp))
-            Text(
-                text = "NEXT SESSIONS",
-                fontWeight = FontWeight(700),
-                fontSize = 16.sp,
-                lineHeight = 19.5.sp,
-                textAlign = TextAlign.Center,
-                color = PRIMARY_BLACK,
-                fontFamily = montserratFamily
-            )
-            Spacer(Modifier.height(24.dp))
-            for (nextClass: UpliftClass in upliftClass?.nextSessions ?: listOf()) {
-                ClassInfoCard(
-                    thisClass = nextClass,
-                    navController = navController,
-                    classDetailViewModel = classDetailViewModel
-                )
-                Spacer(modifier = Modifier.height(12.dp))
-            }
-            Spacer(Modifier.height(36.dp))
-        }
+        NextUpliftClassSessions(
+            upliftClass = upliftClass,
+            classDetailViewModel = classDetailViewModel,
+            navController = navController
+        )
     }
 }
